+++ iOSUbandAPI/Classes/UBandAPITester.swift
@@ -41,7 +41,7 @@ class UBandAPITester: UBandAPIDelegate{
          // do something uBand developer
     }
     
-    func didReceiveBatteryLevelData(uBand: UBandAPI, batteryLevel: Float) {
+    func didReceiveBatteryLevelData(uBand: UBandAPI, batteryLevel: Int) {
          // do something uBand developer
     }
